+++ utils/cicd/main.tf
@@ -18,6 +18,7 @@ module "main" {
   stac_description    = "A STAC API using stac-server"
   stac_server_version = "v5.0.2"
   # stac_server_lambda_zip_filepath
+  # resource_name_suffix
   stac_api_stage                           = "gh"
   stac_api_rootpath                        = "gh"
   stac_api_provisioned_concurrency         = 2